requirements.txt
# list of libraries that are required for the project
tensorflow
pandas
dvc
notebook
numpy 
matplotlib
seaborn 
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors 
-e . 